Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I don't think it's quite right to move just the /cache mounting, as the very next lines manipulate things that are supposed to be within /cache.

How about just change the whole "on init" section to "on fs"?  It's all really about setting up the filesystem, after all.  The only exceptions are the sched changes (do we really need that?) and the apanic stuff (which I think is obsolete).

Patch-set: 1